webex/content: fix GMail div space handling
I use Textern most in GMail, to write essentially all of my email. It
seems that the text in the external editor -> GMail textbox pipeline
gets mangled.

Specifically, I see two phenomena:
- the leading whitespace on each line gets dropped
- adjacent spaces get trimmed to a single space

This seems somewhat GMail-specific. In GitHub, I have no such issues.

The approach taken in this patch is to write to the inner-HTML directly,
in a way that GMail seems to understand.

Closes: #63